+++ flexmeasures/data/schemas/scheduling/storage.py
@@ -15,7 +15,7 @@ from marshmallow.validate import OneOf, ValidationError
 
 from flexmeasures.data.models.time_series import Sensor
 from flexmeasures.data.schemas.units import QuantityField
-from flexmeasures.data.schemas.sensors import QuantityOrSensor, TimeSeriesOrSensor
+from flexmeasures.data.schemas.sensors import VariableQuantityField
 
 from flexmeasures.utils.unit_utils import ur
 
@@ -54,36 +54,62 @@ class StorageFlexModelSchema(Schema):
     You can use StorageScheduler.deserialize_flex_config to get that filled in.
     """
 
-    soc_at_start = fields.Float(required=True, data_key="soc-at-start")
+    soc_at_start = QuantityField(
+        required=True,
+        to_unit="MWh",
+        default_src_unit="dimensionless",  # placeholder, overridden in __init__
+        return_magnitude=True,
+        data_key="soc-at-start",
+    )
 
-    soc_min = fields.Float(validate=validate.Range(min=0), data_key="soc-min")
-    soc_max = fields.Float(data_key="soc-max")
+    soc_min = QuantityField(
+        validate=validate.Range(
+            min=0
+        ),  # change to min=ur.Quantity("0 MWh") in case return_magnitude=False
+        to_unit="MWh",
+        default_src_unit="dimensionless",  # placeholder, overridden in __init__
+        return_magnitude=True,
+        data_key="soc-min",
+    )
+    soc_max = QuantityField(
+        to_unit="MWh",
+        default_src_unit="dimensionless",  # placeholder, overridden in __init__
+        return_magnitude=True,
+        data_key="soc-max",
+    )
 
-    power_capacity_in_mw = QuantityOrSensor(
+    power_capacity_in_mw = VariableQuantityField(
         "MW", required=False, data_key="power-capacity"
     )
 
-    consumption_capacity = QuantityOrSensor(
+    consumption_capacity = VariableQuantityField(
         "MW", data_key="consumption-capacity", required=False
     )
-    production_capacity = QuantityOrSensor(
+    production_capacity = VariableQuantityField(
         "MW", data_key="production-capacity", required=False
     )
 
-    # Timezone placeholder for the soc_maxima, soc_minima and soc_targets fields are overridden in __init__
-    soc_maxima = TimeSeriesOrSensor(
-        unit="MWh", timezone="placeholder", data_key="soc-maxima"
+    # Timezone placeholders for the soc_maxima, soc_minima and soc_targets fields are overridden in __init__
+    soc_maxima = VariableQuantityField(
+        to_unit="MWh",
+        default_src_unit="dimensionless",  # placeholder, overridden in __init__
+        timezone="placeholder",
+        data_key="soc-maxima",
     )
 
-    soc_minima = TimeSeriesOrSensor(
-        unit="MWh",
+    soc_minima = VariableQuantityField(
+        to_unit="MWh",
+        default_src_unit="dimensionless",  # placeholder, overridden in __init__
         timezone="placeholder",
         data_key="soc-minima",
         value_validator=validate.Range(min=0),
     )
 
-    soc_targets = TimeSeriesOrSensor(
-        unit="MWh", timezone="placeholder", data_key="soc-targets"
+    soc_targets = VariableQuantityField(
+        to_unit="MWh",
+        default_src_unit="dimensionless",  # placeholder, overridden in __init__
+        timezone="placeholder",
+        data_key="soc-targets",
     )
 
     soc_unit = fields.Str(
@@ -94,12 +120,13 @@ class StorageFlexModelSchema(Schema):
             ]
         ),
         data_key="soc-unit",
-    )  # todo: allow unit to be set per field, using QuantityField("%", validate=validate.Range(min=0, max=1))
+        required=False,
+    )
 
-    charging_efficiency = QuantityOrSensor(
+    charging_efficiency = VariableQuantityField(
         "%", data_key="charging-efficiency", required=False
     )
-    discharging_efficiency = QuantityOrSensor(
+    discharging_efficiency = VariableQuantityField(
         "%", data_key="discharging-efficiency", required=False
     )
 
@@ -107,14 +134,22 @@ class StorageFlexModelSchema(Schema):
         data_key="roundtrip-efficiency", required=False
     )
 
-    storage_efficiency = QuantityOrSensor(
+    storage_efficiency = VariableQuantityField(
         "%", default_src_unit="dimensionless", data_key="storage-efficiency"
     )
     prefer_charging_sooner = fields.Bool(data_key="prefer-charging-sooner")
 
-    soc_gain = fields.List(QuantityOrSensor("MW"), data_key="soc-gain", required=False)
+    soc_gain = fields.List(
+        VariableQuantityField("MW"),
+        data_key="soc-gain",
+        required=False,
+        validate=validate.Length(min=1),
+    )
     soc_usage = fields.List(
-        QuantityOrSensor("MW"), data_key="soc-usage", required=False
+        VariableQuantityField("MW"),
+        data_key="soc-usage",
+        required=False,
+        validate=validate.Length(min=1),
     )
 
     future_reward = fields.Float(data_key="future-reward", required=False, default=None)
@@ -122,25 +157,51 @@ class StorageFlexModelSchema(Schema):
         data_key="soc-deviation-penalty", required=False, default=None
     )
 
-    def __init__(self, start: datetime, sensor: Sensor, *args, **kwargs):
+    def __init__(
+        self,
+        start: datetime,
+        sensor: Sensor,
+        *args,
+        default_soc_unit: str | None = None,
+        **kwargs,
+    ):
         """Pass the schedule's start, so we can use it to validate soc-target datetimes."""
         self.start = start
         self.sensor = sensor
-        self.soc_maxima = TimeSeriesOrSensor(
-            unit="MWh", timezone=sensor.timezone, data_key="soc-maxima"
+
+        # guess default soc-unit
+        if default_soc_unit is None:
+            if self.sensor.unit in ("MWh", "kWh"):
+                default_soc_unit = self.sensor.unit
+            elif self.sensor.unit in ("MW", "kW"):
+                default_soc_unit = self.sensor.unit + "h"
+
+        self.soc_maxima = VariableQuantityField(
+            to_unit="MWh",
+            default_src_unit=default_soc_unit,
+            timezone=sensor.timezone,
+            data_key="soc-maxima",
         )
 
-        self.soc_minima = TimeSeriesOrSensor(
-            unit="MWh",
+        self.soc_minima = VariableQuantityField(
+            to_unit="MWh",
+            default_src_unit=default_soc_unit,
             timezone=sensor.timezone,
             data_key="soc-minima",
             value_validator=validate.Range(min=0),
         )
-        self.soc_targets = TimeSeriesOrSensor(
-            unit="MWh", timezone=sensor.timezone, data_key="soc-targets"
+        self.soc_targets = VariableQuantityField(
+            to_unit="MWh",
+            default_src_unit=default_soc_unit,
+            timezone=sensor.timezone,
+            data_key="soc-targets",
         )
 
         super().__init__(*args, **kwargs)
+        if default_soc_unit is not None:
+            for field in self.fields.keys():
+                if field.startswith("soc_"):
+                    setattr(self.fields[field], "default_src_unit", default_soc_unit)
 
     @validates_schema
     def check_whether_targets_exceed_max_planning_horizon(self, data: dict, **kwargs):
@@ -191,33 +252,8 @@ class StorageFlexModelSchema(Schema):
     @post_load
     def post_load_sequence(self, data: dict, **kwargs) -> dict:
         """Perform some checks and corrections after we loaded."""
-        # currently we only handle MWh internally
-        # TODO: review when we moved away from capacity having to be described in MWh
-        if data.get("soc_unit") == "kWh":
-            data["soc_at_start"] /= 1000.0
-            if data.get("soc_min") is not None:
-                data["soc_min"] /= 1000.0
-            if data.get("soc_max") is not None:
-                data["soc_max"] /= 1000.0
-            if (
-                not isinstance(data.get("soc_targets"), Sensor)
-                and data.get("soc_targets") is not None
-            ):
-                for target in data["soc_targets"]:
-                    target["value"] /= 1000.0
-            if (
-                not isinstance(data.get("soc_minima"), Sensor)
-                and data.get("soc_minima") is not None
-            ):
-                for minimum in data["soc_minima"]:
-                    minimum["value"] /= 1000.0
-            if (
-                not isinstance(data.get("soc_maxima"), Sensor)
-                and data.get("soc_maxima") is not None
-            ):
-                for maximum in data["soc_maxima"]:
-                    maximum["value"] /= 1000.0
-            data["soc_unit"] = "MWh"
+        # currently we only handle MWh internally, and the conversion to MWh happened during deserialization
+        data["soc_unit"] = "MWh"
 
         # Convert efficiency to dimensionless (to the (0,1] range)
         if data.get("roundtrip_efficiency") is not None: